hotel-management/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
pydantic